+++ conf/app.routes
@@ -1,5 +1,3 @@
 # microservice specific routes
 
-
-GET        /:nino/:taxYear                  uk.gov.hmrc.taxhistory.controllers.EmploymentHistoryController.getEmploymentHistory(nino:String,taxYear:Int)
 GET        /:nino/:taxYear/employments      uk.gov.hmrc.taxhistory.controllers.EmploymentController.getEmployments(nino:String,taxYear:Int)
