+++ test/uk/gov/hmrc/taxhistory/services/EmploymentServiceSpec.scala
@@ -211,7 +211,7 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
     "return any non success status response from get Nps Employments api" in {
       when(mockNpsConnector.getEmployments(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
         .thenReturn(Future.successful(HttpResponse(BAD_REQUEST, Some(npsEmploymentResponse))))
-      val response =  await(TestEmploymentService.getEmploymentHistory(testNino.toString(),2016))
+      val response =  await(TestEmploymentService.getEmployments(testNino.toString(),2016))
       response mustBe a[HttpResponse]
       response.status mustBe BAD_REQUEST
     }
@@ -219,7 +219,7 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
     "return not found status response from get Nps Employments api" in {
       when(mockNpsConnector.getEmployments(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
         .thenReturn(Future.successful(HttpResponse(OK, Some(JsArray(Seq.empty)))))
-      val response =  await(TestEmploymentService.getEmploymentHistory(testNino.toString(),2016))
+      val response =  await(TestEmploymentService.getEmployments(testNino.toString(),2016))
       response mustBe a[HttpResponse]
       response.status mustBe NOT_FOUND
     }
@@ -231,7 +231,7 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
         .thenReturn(Future.successful(HttpResponse(BAD_REQUEST, Some(rtiEmploymentResponse))))
       when(mockNpsConnector.getIabds(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
         .thenReturn(Future.successful(HttpResponse(BAD_REQUEST, Some(iabdsJsonResponse))))
-      val response =  await(TestEmploymentService.getEmploymentHistory(testNino.toString(),2016))
+      val response =  await(TestEmploymentService.getEmployments(testNino.toString(),2016))
       response mustBe a[HttpResponse]
       response.status mustBe OK
     }
@@ -243,19 +243,19 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
         .thenReturn(Future.successful(HttpResponse(OK, Some(iabdsJsonResponse))))
       when(mockRtiDataConnector.getRTIEmployments(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
         .thenReturn(Future.successful(HttpResponse(OK, Some(rtiEmploymentResponse))))
-      val response =  await(TestEmploymentService.getEmploymentHistory(testNino.toString(),2016))
+      val response =  await(TestEmploymentService.getEmployments(testNino.toString(),2016))
       response mustBe a[HttpResponse]
       response.status mustBe OK
-      val payAsYouEarnDetails = response.json.as[PayAsYouEarnDetails]
-      val employments = payAsYouEarnDetails.employments
+      val employments = response.json.as[List[uk.gov.hmrc.taxhistory.model.api.Employment]]
+
       employments.size mustBe 1
       employments.head.employerName mustBe "Aldi"
       employments.head.payeReference mustBe "531/J4816"
-      employments.head.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
-      employments.head.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
-      employments.head.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
+     // employments.head.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
+     // employments.head.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
+     // employments.head.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
       employments.head.startDate mustBe startDate
-      employments.head.companyBenefits mustBe List(CompanyBenefit("Car Fuel Benefit",100,"CarFuelBenefit"), CompanyBenefit("Van Benefit",100,"VanBenefit"))
+     // employments.head.companyBenefits mustBe List(CompanyBenefit("Car Fuel Benefit",100,"CarFuelBenefit"), CompanyBenefit("Van Benefit",100,"VanBenefit"))
       employments.head.endDate mustBe None
     }
 
@@ -263,19 +263,17 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
       val rtiData = rtiEmploymentResponse.as[RtiData]
       val npsEmployments = npsEmploymentResponse.as[List[NpsEmployment]]
 
-      val response = await(TestEmploymentService.getPayAsYouEarnDetails(testNino, TaxYear(2016))(npsEmployments))
+      val response = await(TestEmploymentService.mergeAndRetrieveEmployments(testNino, TaxYear(2016))(npsEmployments))
       response mustBe a[HttpResponse]
 
       response.status mustBe OK
-      val payAsYouEarnDetails = response.json.as[PayAsYouEarnDetails]
-
-      val employments = payAsYouEarnDetails.employments
+      val employments = response.json.as[List[uk.gov.hmrc.taxhistory.model.api.Employment]]
 
       employments.head.employerName mustBe "Aldi"
       employments.head.payeReference mustBe "531/J4816"
-      employments.head.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
-      employments.head.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
-      employments.head.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
+     // employments.head.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
+     // employments.head.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
+     // employments.head.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
       employments.head.startDate mustBe startDate
       employments.head.endDate mustBe None
     }
@@ -288,13 +286,12 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
           .thenReturn(Future.successful(HttpResponse(OK, Some(iabdsJsonResponse))))
         when(mockRtiDataConnector.getRTIEmployments(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
           .thenReturn(Future.successful(HttpResponse(OK, Some(rtiEmploymentResponse))))
-        val response = await(TestEmploymentService.getEmploymentHistory(testNino.toString(), 2016))
+        val response = await(TestEmploymentService.getEmployments(testNino.toString(), 2016))
         response mustBe a[HttpResponse]
         response.status mustBe OK
-        val payAsYouEarnDetails = response.json.as[PayAsYouEarnDetails]
-        val employments = payAsYouEarnDetails.employments
+        val employments = response.json.as[List[uk.gov.hmrc.taxhistory.model.api.Employment]]
         employments.size mustBe 1
-        payAsYouEarnDetails.allowances mustBe List(Allowance("Flat Rate Job Expenses",200, "FlatRateJobExpenses"))
+        //payAsYouEarnDetails.allowances mustBe List(Allowance("Flat Rate Job Expenses",200, "FlatRateJobExpenses"))
       }
 
       "nps receivingJobseekersAllowance and otherIncomeSourceIndicator is true form list of employments" in {
@@ -304,7 +301,7 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
           .thenReturn(Future.successful(HttpResponse(OK, Some(iabdsJsonResponse))))
         when(mockRtiDataConnector.getRTIEmployments(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
           .thenReturn(Future.successful(HttpResponse(OK, Some(rtiEmploymentResponse))))
-        val response = await(TestEmploymentService.getEmploymentHistory(testNino.toString(), 2016))
+        val response = await(TestEmploymentService.getEmployments(testNino.toString(), 2016))
         response mustBe a[HttpResponse]
         response.status mustBe NOT_FOUND
       }
@@ -318,7 +315,7 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
           .thenReturn(Future.successful(HttpResponse(OK, Some(iabdsJsonResponse))))
         when(mockRtiDataConnector.getRTIEmployments(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
           .thenReturn(Future.successful(HttpResponse(OK, Some(rtiEmploymentResponse))))
-        val response = await(TestEmploymentService.getEmploymentHistory(testNino.toString(), 2016))
+        val response = await(TestEmploymentService.getEmployments(testNino.toString(), 2016))
         response mustBe a[HttpResponse]
         response.status mustBe NOT_FOUND
       }
@@ -330,7 +327,7 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
           .thenReturn(Future.successful(HttpResponse(OK, Some(iabdsJsonResponse))))
         when(mockRtiDataConnector.getRTIEmployments(Matchers.any(), Matchers.any())(Matchers.any[HeaderCarrier]))
           .thenReturn(Future.successful(HttpResponse(OK, Some(rtiEmploymentResponse))))
-        val response = await(TestEmploymentService.getEmploymentHistory(testNino.toString(), 2016))
+        val response = await(TestEmploymentService.getEmployments(testNino.toString(), 2016))
         response mustBe a[HttpResponse]
         response.status mustBe NOT_FOUND
       }
@@ -428,12 +425,12 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
       val employment=TestEmploymentService.buildEmployment(Some(rtiData.employments),Some(iabds), npsEmployments.head)
       employment.employerName mustBe "Aldi"
       employment.payeReference mustBe "0531/J4816"
-      employment.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
-      employment.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
-      employment.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
+      //employment.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
+      //employment.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
+      //employment.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
       employment.startDate mustBe startDate
       employment.endDate mustBe None
-      employment.companyBenefits.size mustBe  8
+      //employment.companyBenefits.size mustBe  8
     }
 
     "Build employment1 when there is no  data for rti and Iabd" in {
@@ -444,10 +441,10 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
       val employment=TestEmploymentService.buildEmployment(None,None, npsEmployments.head)
       employment.employerName mustBe "Aldi"
       employment.payeReference mustBe "0531/J4816"
-      employment.taxablePayTotal mustBe None
-      employment.taxTotal mustBe None
-      employment.earlierYearUpdates mustBe List()
-      employment.companyBenefits.size mustBe  0
+     // employment.taxablePayTotal mustBe None
+     // employment.taxTotal mustBe None
+     // employment.earlierYearUpdates mustBe List()
+     // employment.companyBenefits.size mustBe  0
       employment.startDate mustBe startDate
       employment.endDate mustBe None
     }
@@ -459,10 +456,10 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
       val employment=TestEmploymentService.buildEmployment(None,Some(iabds), npsEmployments.head)
       employment.employerName mustBe "Aldi"
       employment.payeReference mustBe "0531/J4816"
-      employment.taxablePayTotal mustBe None
-      employment.taxTotal mustBe None
-      employment.earlierYearUpdates mustBe List()
-      employment.companyBenefits.size mustBe  8
+     // employment.taxablePayTotal mustBe None
+     // employment.taxTotal mustBe None
+     // employment.earlierYearUpdates mustBe List()
+     // employment.companyBenefits.size mustBe  8
       employment.startDate mustBe startDate
       employment.endDate mustBe None
     }
@@ -475,10 +472,10 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
       val employment=TestEmploymentService.buildEmployment(None,Some(iabds), npsEmployments.head)
       employment.employerName mustBe "Aldi"
       employment.payeReference mustBe "0531/J4816"
-      employment.taxablePayTotal mustBe None
-      employment.taxTotal mustBe None
-      employment.earlierYearUpdates mustBe List()
-      employment.companyBenefits.size mustBe  8
+     // employment.taxablePayTotal mustBe None
+     // employment.taxTotal mustBe None
+     // employment.earlierYearUpdates mustBe List()
+     // employment.companyBenefits.size mustBe  8
       employment.startDate mustBe startDate
       employment.endDate mustBe None
     }
@@ -490,10 +487,10 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
       val employment=TestEmploymentService.buildEmployment(Some(rtiData.employments),None, npsEmployments.head)
       employment.employerName mustBe "Aldi"
       employment.payeReference mustBe "0531/J4816"
-      employment.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
-      employment.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
-      employment.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
-      employment.companyBenefits.size mustBe  0
+     // employment.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
+     // employment.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
+     // employment.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
+     // employment.companyBenefits.size mustBe  0
       employment.startDate mustBe startDate
       employment.endDate mustBe None
     }
@@ -505,10 +502,10 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
       val employment=TestEmploymentService.buildEmployment(Some(rtiData.employments),Some(Nil), npsEmployments.head)
       employment.employerName mustBe "Aldi"
       employment.payeReference mustBe "0531/J4816"
-      employment.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
-      employment.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
-      employment.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
-      employment.companyBenefits.size mustBe  0
+     // employment.taxablePayTotal mustBe Some(BigDecimal.valueOf(20000.00))
+     // employment.taxTotal mustBe Some(BigDecimal.valueOf(1880.00))
+     // employment.earlierYearUpdates mustBe List(EarlierYearUpdate(-600.99,-10.99,LocalDate.parse("2016-06-01")))
+     // employment.companyBenefits.size mustBe  0
       employment.startDate mustBe startDate
       employment.endDate mustBe None
     }
@@ -577,13 +574,6 @@ class EmploymentServiceSpec extends PlaySpec with MockitoSugar with TestUtil{
 
       val onlyRtiEmployments = TestEmploymentService.onlyInRTI(rtiEmployments,npsEmployments)
       onlyRtiEmployments.size mustBe 0
-    }
-
-    "return api employments model when get employments is called" in {
-      val response =  await(TestEmploymentService.getEmployments(testNino.toString(),2016))
-      response mustBe a[HttpResponse]
-      response.status mustBe OK
-      response.json mustBe employmentsApiJsonResponse
     }
   }
